requirements.txt
crewai>=0.28.1
google-generativeai>=0.3.1
matplotlib>=3.7.1
plotly>=5.18.0
gradio>=4.19.1
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
langchain>=0.0.312
langchain-core>=0.1.15
langchain-community>=0.0.12
numpy>=1.24.0 